% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smart.R
\name{read_smart}
\alias{read_smart}
\alias{read_rda}
\title{A smart function for reading data files}
\usage{
read_smart(x, ...)

read_rda(x)
}
\arguments{
\item{x}{Name of file to read.}

\item{...}{Other args passed on to read-in function, see details for
more info.}
}
\value{
Info from file returned as R data object.
}
\description{
Reads data based on file extension

Reads in .rda files like any other data file would (outputs an object, which you can name whatever you want)
}
\details{
Here is the function logic \itemize{
  \item csv = readr::read_csv
  \item rds = readRDS
  \item rda|Rdata = load
  \item xlsx = openxlsx::read.xlsx
  \item sav = haven::read_sav
  \item dta = haven::read_dta
  \item tsv = read.delim
  \item fst = fst::read_fst
  \item dt|data.table = data.table::fred
  \item txt|table = read.table
  \item html = xml2::read_html
  \item xml = xml2::read_xml
  \item json = jsonlite::stream_in
  \item yaml = yaml::read_yaml
}
}
\examples{
## save congress data as .rds, .csv, and .rda files
saveRDS(congress, "congress.rds")
save(congress, file = "congress.rda")
write.csv(congress, "congress.csv", row.names = FALSE)

## now read and view each file
(c1 <- read_smart("congress.rds"))
(c2 <- read_smart("congress.csv"))
(c3 <- read_smart("congress.rda"))

}
